# docs/swagger.yaml
basePath: /
definitions:
  views.AuthRequest:
    properties:
      email:
        type: string
      login:
        type: string
      password:
        type: string
    type: object
  views.File2DBResponse:
    additionalProperties: true
    type: object
  views.GenerateMDRequest:
    properties:
      content:
        example: Текст документа, который нужно законспектировать
        type: string
    type: object
  views.GenerateTasksRequest:
    properties:
      content:
        example: Контекст и/или промт для генерации заданий
        type: string
    type: object
  views.MarkdownResponse:
    properties:
      markdown:
        example: '# Конспект ...'
        type: string
    type: object
  views.SWGError:
    properties:
      error:
        example: error
        type: string
    type: object
  views.SWGMessage:
    properties:
      message:
        example: some info
        type: string
    type: object
  views.TasksMarkdownResponse:
    properties:
      markdown:
        example: |-
          1. Вопрос...
             A) ...
             B) ...
        type: string
    type: object
  views.Tokens:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  views.TranscribeResponse:
    properties:
      duration_seconds:
        example: 123.45
        type: number
      filename:
        example: lecture.mp3
        type: string
      language:
        example: ru
        type: string
      model:
        example: nova-2-general
        type: string
      text:
        example: полная расшифровка аудио
        type: string
    type: object
  views.UserRegister:
    properties:
      email:
        type: string
      login:
        type: string
      pw1:
        type: string
      pw2:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: help@breezynotes.ru
    name: Alex "bustard" Provor
    url: https://contacts.breezynotes.ru
  description: yoyo
  termsOfService: https://about.breezynotes.ru/
  title: flicker rest api
  version: 0.1-.-infDev
paths:
  /api/ai/file2db:
    post:
      consumes:
      - multipart/form-data
      description: Принимает файл, отправляет его в n8n webhook file2db, который сохраняет
        данные во векторную БД
      parameters:
      - description: File to index
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.File2DBResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.SWGError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/views.SWGError'
      summary: Upload file to vector DB
      tags:
      - ai
  /api/ai/file2dbtest:
    post:
      consumes:
      - multipart/form-data
      description: Принимает файл, отправляет его в n8n webhook file2db, который сохраняет
        данные во векторную БД
      parameters:
      - description: File to index
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.File2DBResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.SWGError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/views.SWGError'
      summary: Upload file to vector DB
      tags:
      - ai
  /api/ai/generatemd:
    post:
      consumes:
      - application/json
      description: Принимает текст и отправляет его в n8n webhook, который генерирует
        Markdown-конспект через LLM
      parameters:
      - description: Text content to summarize
        in: body
        name: Content
        required: true
        schema:
          $ref: '#/definitions/views.GenerateMDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.MarkdownResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.SWGError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/views.SWGError'
      summary: Generate Markdown summary
      tags:
      - ai
  /api/ai/generatemd-test:
    post:
      consumes:
      - application/json
      description: Принимает контекст/промт и отправляет его в n8n webhook, который
        генерирует задания (тесты, вопросы) в формате Markdown на основе этого контекста
      parameters:
      - description: Context and/or prompt for tasks generation
        in: body
        name: Content
        required: true
        schema:
          $ref: '#/definitions/views.GenerateTasksRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.TasksMarkdownResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.SWGError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/views.SWGError'
      summary: Generate tasks in Markdown
      tags:
      - ai
  /api/ai/transcribe:
    post:
      consumes:
      - multipart/form-data
      description: Принимает аудио-файл, отправляет его в сервис транскрипции и возвращает
        текст
      parameters:
      - description: Audio file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.TranscribeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.SWGError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/views.SWGError'
      summary: Transcribe audio file
      tags:
      - ai
  /api/auth:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns access/refresh tokens
      parameters:
      - description: Login or Email and Password
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/views.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.Tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.SWGError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/views.SWGError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/views.SWGError'
      summary: Authorize user
      tags:
      - auth
  /api/auth/reg:
    post:
      consumes:
      - application/json
      description: Validates registration data, creates user and returns tokens
      parameters:
      - description: Reg data
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/views.UserRegister'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.SWGError'
        "302":
          description: Found
          schema:
            $ref: '#/definitions/views.SWGError'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.SWGError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/views.SWGError'
      summary: Register new user
      tags:
      - auth
  /api/auth/token:
    get:
      description: Checks access token from cookie, tries to refresh if expired. If
        410 (GONE) need to re-auth
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.SWGMessage'
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.SWGError'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/views.SWGError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/views.SWGError'
      summary: Validate token (uses cookies)
      tags:
      - auth
  /api/health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.SWGMessage'
      summary: check health of gateway
      tags:
      - healthz
schemes:
- http
swagger: "2.0"